Calculate retention

+++ src/main/java/com/vigen/plexonic/analytics/api/presentation/RetentionController.java
@@ -21,7 +21,7 @@ public class RetentionController {
             params = {"type", "date"}
     )
     public @ResponseBody
-    Integer calculateRetention(
+    double calculateRetention(
             @RequestParam("type") RetentionType type,
             @RequestParam("date") @DateTimeFormat(pattern = "dd/MM/yyyy") Date date) {
         return retentionService.calculateRetention(type, date);
